third_party/WebKit/LayoutTests/platform/win/external/wpt/html/semantics/scripting-1/the-script-element/module/crossorigin-expected.txt

This is a testharness.js-based test.
FAIL Root module, Error in CORS-different-origin script assert_equals: Unexpected _log value expected "running,0-0" but got "0-0"
FAIL Root module, Error in CORS-same-origin script assert_equals: Unexpected _log value expected "running,8-1" but got "1-1"
PASS Root module, Blocked script download, missing CORS ACAO header 
PASS Root module, Blocked script download, mismatched CORS ACAO header 
FAIL Imported module, Error in CORS-different-origin script assert_equals: Unexpected _log value expected "running,0-0" but got "11-9"
FAIL Imported module, Error in CORS-same-origin script assert_equals: Unexpected _log value expected "running,8-1" but got "11-9"
FAIL Imported module, Blocked script download, missing CORS ACAO header assert_equals: Unexpected _log value expected "error" but got "11-9"
FAIL Imported module, Blocked script download, mismatched CORS ACAO header assert_equals: Unexpected _log value expected "error" but got "11-9"
Harness: the test ran to completion.
